LayoutTests/platform/gtk/accessibility/media-emits-object-replacement.html

<!DOCTYPE HTML>
<html>
<head>
<script>
function buildAccessibilityTree(accessibilityObject, indent) {
    var str = "";
    for (var i = 0; i < indent; i++)
        str += "    ";
    str += accessibilityObject.role;
    str += " " + accessibilityObject.stringValue;
    str += "\n";
    document.getElementById("tree").innerText += str;

    if (accessibilityObject.stringValue.indexOf('End of test') >= 0)
        return false;

    var count = accessibilityObject.childrenCount;
    for (var i = 0; i < count; ++i) {
        if (!buildAccessibilityTree(accessibilityObject.childAtIndex(i), indent + 1))
            return false;
    }

    return true;
}
</script>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>

<div id="text1">
a <video width=100 height=100 controls></video> b
</div>

<br>

<div id="text2">
a <audio width=100 height=100 controls></audio> b
</div>

</div>

<div>End of test</div>

<pre id="tree"></pre>

<p id="description"></p>
<div id="console"></div>

<script>
description("This tests ensures that if video or audio tags are used, they will emit an object replacement character in a range for string operation.")

if (window.accessibilityController) {
    // Build the accessibility tree up until 'End of test' is encountered.
    document.body.focus();
    buildAccessibilityTree(accessibilityController.focusedElement, 0);
}
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
